feat(AHWR-1047): Adding tests for review and follow-up claims with visit date before MH release date
