+++ test/specs/postMH/test.pre.mh.journeys.js
@@ -1,44 +1,32 @@
 import { expect, browser, $ } from "@wdio/globals";
 import {
-  getDevSignInUrl,
-  fillAndSubmitSBI,
-  clickSubmitButton,
   clickOnElementAndContinue,
-  fillInputAndContinue,
-  verifySubmission,
   clickStartNewClaimButton,
+  clickSubmitButton,
   enterPreMHReleaseDateAndContinue,
-  enterWhenTestingWasCarriedOutAndContinue,
-  selectSheepTestsAndContinue,
-  createSheepReviewClaim,
+  fillAndSubmitSBI,
+  getDevSignInUrl,
 } from "../../utils/common.js";
 import {
-  NUMBER_OF_ANIMALS_TESTED,
-  VETS_NAME,
-  VET_RCVS_NUMBER,
-  SUBMIT_CLAIM_BUTTON,
-  REFERENCE,
+  createPreMultipleHerdPigsFollowUp,
+  createPreMultipleHerdSheepFollowUp,
+} from "../../utils/follow-up-claim.js";
+import { createPigsReviewClaim, createSheepReviewClaim } from "../../utils/review-claim.js";
+import {
   CLAIMS_MAIN_HEADING_SELECTOR,
   EXTERNAL_GOV_LINK,
   getTypeOfLivestockSelector,
   getTypeOfReviewSelector,
-  getSpeciesNumbersSelector,
   getConfirmCheckDetailsSelector,
-  getSheepEndemicsPackageSelector,
-  getTestResultSelector,
 } from "../../utils/selectors.js";
-import {
-  PRE_MULTIPLE_HERDS_SBI,
-  PRE_MULTIPLE_HERDS_SHEEP_AGREEMENT_REF,
-  JOHNES_DISEASE,
-} from "../../utils/constants.js";
+import { PRE_MULTIPLE_HERD_SBI, PRE_MULTIPLE_HERD_AGREEMENT_REF } from "../../utils/constants.js";
 import { approveClaim } from "../../utils/backoffice-common.js";
 
-describe("Pre-MH journeys for sheep when MH is switched on", () => {
+describe("Pre-MH journeys when MH is switched on", () => {
   it("cannot create a second review claim for sheep species when visit date is before mh release date and within 10 months of its pre-MH review claim", async () => {
     // Using an SBI that already has a pre-mh review claim
     await browser.url(getDevSignInUrl("claim"));
-    await fillAndSubmitSBI(PRE_MULTIPLE_HERDS_SBI);
+    await fillAndSubmitSBI(PRE_MULTIPLE_HERD_SBI);
     await $(getConfirmCheckDetailsSelector("yes")).click();
     await clickSubmitButton();
     await clickStartNewClaimButton();
@@ -55,36 +43,27 @@ describe("Pre-MH journeys for sheep when MH is switched on", () => {
   });
 
   it("can create a follow-up claim for a pre-MH sheep review claim if the follow-up visit date is before the MH release date", async () => {
-    const claimNumber = await createSheepReviewClaim(PRE_MULTIPLE_HERDS_SBI, {
+    const claimNumber = await createSheepReviewClaim(PRE_MULTIPLE_HERD_SBI, {
       multipleHerdFlag: true,
       isUnnamedHerdClaimPresent: true,
     });
 
     // Approve the mh review claim
-    await approveClaim(PRE_MULTIPLE_HERDS_SHEEP_AGREEMENT_REF, claimNumber);
+    await approveClaim(PRE_MULTIPLE_HERD_AGREEMENT_REF, claimNumber);
 
-    // now doing a follow-up for the pre-mh claim by using a visit date before the mh release date
-    await browser.url(getDevSignInUrl("claim"));
-    await fillAndSubmitSBI(PRE_MULTIPLE_HERDS_SBI);
-    await $(getConfirmCheckDetailsSelector("yes")).click();
-    await clickSubmitButton();
-    await clickStartNewClaimButton();
-    await clickOnElementAndContinue(getTypeOfLivestockSelector("sheep"));
-    await clickOnElementAndContinue(getTypeOfReviewSelector("endemics"));
+    // now doing a follow-up for the pre-mh sheep review claim by using a visit date before the mh release date
+    await createPreMultipleHerdSheepFollowUp(PRE_MULTIPLE_HERD_SBI);
+  });
 
-    await enterPreMHReleaseDateAndContinue();
+  it("can create a review and its follow-up claim for pigs if the visit date is before the MH release date", async () => {
+    const claimNumber = await createPigsReviewClaim(PRE_MULTIPLE_HERD_SBI, {
+      enterVisitDateAndContinueFunc: enterPreMHReleaseDateAndContinue,
+    });
 
-    await enterWhenTestingWasCarriedOutAndContinue("whenTheVetVisitedTheFarmToCarryOutTheReview");
-    await clickOnElementAndContinue(getSpeciesNumbersSelector("yes"));
-    await fillInputAndContinue(NUMBER_OF_ANIMALS_TESTED, "10");
-    await fillInputAndContinue(VETS_NAME, "Mr Auto Test");
-    await fillInputAndContinue(VET_RCVS_NUMBER, "1234567");
-    await clickOnElementAndContinue(getSheepEndemicsPackageSelector("improvedEwePerformance"));
-    await selectSheepTestsAndContinue([JOHNES_DISEASE]);
-    await clickOnElementAndContinue(getTestResultSelector("positive"));
-    await $(SUBMIT_CLAIM_BUTTON).click();
-    await verifySubmission("Claim complete");
+    // Approve the mh review claim
+    await approveClaim(PRE_MULTIPLE_HERD_AGREEMENT_REF, claimNumber);
 
-    await expect($(REFERENCE)).toHaveText(expect.stringContaining("FUSH"));
+    // now doing a follow-up for the pre-mh pigs review claim by using a visit date before the mh release date
+    await createPreMultipleHerdPigsFollowUp(PRE_MULTIPLE_HERD_SBI);
   });
 });
