+++ test/specs/test.claim-compliance.journeys.js
@@ -1,6 +1,7 @@
 import { browser } from "@wdio/globals";
-import { createSheepReviewClaim, getDevSignInUrl } from "../utils/common.js";
+import { getDevSignInUrl } from "../utils/common.js";
 import { assertClaimToBeInCheck, assertClaimToBeOnHold } from "../utils/common-assertions.js";
+import { createSheepReviewClaim } from "../utils/review-claim.js";
 
 const fillerSbis = ["106416234", "107361798", "107645299", "106258541", "107346082"];
