+++ wdio.conf.js
@@ -161,7 +161,7 @@ export const config = {
   // See the full list at http://mochajs.org/
   mochaOpts: {
     ui: "bdd",
-    timeout: 120000,
+    timeout: 180000,
   },
 
   //
